Remove TINT field in limits struct
Having this field was a problem because it made impossible
to take a value of the union without checking the type
which generates this limits. This complexity is not needed,
because TUINT can cover the maximum of any integer type,
and in the case of the min, we only have to negate the value.