Return typedef symbols in struct namespaces
Typedef symbols are stored in NS_IDEN namespace, which is the
default namespace of the lexer. When the definition of struct
type is done, the lexer take the namespace of the struct, and
it means that typedef symbols are not seen in such context,
making impossible to declara fields with typedefed types.