Simplify definitions in the IR
Having the definition of initializers in the nex line was causing
that was impossible to be able to emit variables until parsing
the next element, which forced us to deal with a lastsym variable
and several cases. It is better to know in the same line if
the symbol is initialized or not.