[cc2] Allocate space for symbols always
The current code was passing 1 in alloc parameter of defsym()
only when there was not an initializer, but this semantic
was wrong and it creates problems with qbe. The correct
semantic is to pass 1 and let to the function to determine
what it has to do based in the information of the symbol.