[cc2] Add skeleton for code.c
This file is going to have the functions that implement the common
functions about the 3 address instructions. This code is basically
imported from cc2.old, but this time the basic structure of the
instructions is a 3 address, instead of a 2 address. The original
cc2 had 2 address instructions because it was only z80 targered,
but this new version wants to be multi targered, so it is better
a more general approach, although it will waste memory in accumulator
based architectures.