[cc2] Rename addressability() to apply()
Addressability() was basically running over the nodes of the current
function and calling a function for every statement. This is a general
function and it is better to pass it the function we can call for
every statement.
Since the function is now a generic function, the best place for it
is in node.c