[cc1 - cc2] Output function return type in cc1 and cc2
Print function return type in cc1 and cc2.
In cc1 the return type is printed before the symbol type 'F', so that in cc2
we can first pop 'F' from the stack and see if we need to pop the return type.

The return type needs to be stored somewhere in cc2, so I added it
to Symbol as Symbol.rtype.