[cc1] Fix use-affer-free bug in switches
Switches build a list of cases which were emitted after the
end of the switch statement, and since the switch statement is
usually a compound statement, it meant that the symbols created
in the cases were freed when they were emitted. To solve this
problem this patch simplifies everything emitting the cases
on the fly, letting the job of creating the switch table to the
backend.