[cc2] Move optm to target code
All the independent optimizations are done in cc1, so there is no much room
for target independent optimizations in cc2. At this moment these target
independent optimizations were done mainly because qbe needs them, but
we there are some changes that qbe needs that are not going to be
useful for other targets (and in some cases even they are going to be
worse).