[cc2-qbe] Ensure correctness of basic blocks
We have to take care of how basic blocks are built in QBE,
so we have to ensure that all the basic blocks of the program
are finalized with a jump of any type. We also need a label
in the next statement of a branch, because in branches in QBE
we have to use one label for true and another one for false.