[cc2-qbe] Add OBRANCH generation
OBRANCH is similar to OJMP, but it has a condition expression
and two labels instead of only one. A new function label()
is created to convert labels into nodes, which are needed
for jumps. Before this commit the code was making tricks
to link labels and nodes. This is not the best solution,
but it is the simpler.