[cc2-qbe] Remove l and r variables in cgen()
These variables are not useful anymore after the last change
to load, because we don't keep them updated. The actuar left
and right are in np.
This way of working is generating some "memory leaks", because
some nodes become unlinked from the root, and it means that
they will not be freed until cleaning the arena.