[cc1] Remove undefined behaviour in specifier
Variables local to loops are created and destroy in every iteration
of the loop, and it means that they (logically) does not retain
the value from the previous iteration. In the case of long long
we were using the value of the previous iteration (the iteration
of the first long), and it was working because moderm compilers
does not create/destroy the variables in this case. It was possible
to create strange results with something like:

	long int long

because in this case p was pointing to type and not to the size.
This patch fixes the problem setting the value of p to NULL in
every iteration and explicitily setting the value of p in the case
of long long. If the value of p is not set to the correct value
we will have a segmentation fault and e will discover the error
as soon as possible.