[cc1] Add equiv parameter to eqtype()
Eqtype() compare two types, but it was checking always that
the types were exactly the same, and in some situations is
needed to check that two types are equivalent. Both
situations shared almost all the code, so the best option
is to add a new parameter to control what behaviour
the caller expects.