[cc1] handle correctly arrays in address()
When a address operator is applied to an array then we do not
want to decay the array, because we want a pointer to the array
itself, not to the first element of the array. In the same way,
an array is not a lvalue, but it is legal to take a pointer to it.