[cc1] Do not warn about empty declarations in prototypes
It is fine to have prototypes where parameters lack of a name,
but the current form of the code was giving a warning about
empty declarations. With this patch we still keep the extension
of allowing anonymous parameters (the reallity is that it is
eassier to keep this extension that fordib it :P).