[cc1] Fix pre increment/decrement for pointers
Pre increment/decrement for pointers was broken, mainly
because pointers are not arithmetic types, and the size
of the pointers were calculated in two different places.
Another problem was related to the fact that post inc/dec
generates an op but pre inc/dec generates another one,
and the tests were checking only against the values
generated for post operators.