Remove multi targets in Makefile
This way of building was crappy because it had to clean
all the files in every step, and it was very easy to merge
files from different architectures. After the last changes
the common objects are totally independent of the architecture
so they can be used without problems.

The default target build all the architectures instead of building
only the one controlled by the ARCH variable.