[cc1] Remove lazy printout of types
Types were printed when was needed, when some variable related
to it was printed. This was a really bad idea because it was
generating a disorder in how the types were printed, and in
some cases global types were printed in the body of a function,
being deleted in flushtypes().