[cc1] Make symbol table an ordered hash table again
Since cpp symbols were allocated in the main hash table we had to deal
with symbols inserted out of order, because cpp symbols should stay
in the table forever. This is not needed anymore since we are inserting
cpp symbols in a different table we can get rid of this code.