[cc1] Remove popctx() from compound()
After the commit fea243a this popctx() it is not longer
needed to happen in compound(), where it was a bit difficult
to explain why it was needed, and it was very, very tricky.
A better place where to pop the parameter context is in
decl(), where it already pop the parameter context in the
case of prototypes.