[cc1] Fix inferred array sizes
It seems most everything was in place for this, but mktype was
unconditionally defining the incomplete array type by passing 1 to
newtype. Instead, make newtype use t.prop & TDEFINED to decide whether
to call deftype.

Also, call deftype in initlist to compute the size and emit the type IR.

This fixes 0096-inferredarraysize.c.