[cc1] Fix comparisions
Comparisions are a bit different to other binary oeprators, because
they have two types, the type in which the operation is done and
the type of the result of the operation. This cannot be represented
with out IR because each node strictly has only one type, so the best
solution is to keep the result of the operation in the same type than
the oepration and add a cast after that. cc2 can cover perfctly with
this combination of nodes without any work.