[cc1] Move concatenation of strings to primary()
The concatenation of strings was done in lex.c:/^string
and it was to soon. There was a trick in string() which
look ahead to see if the next character was a ", and then
concatenate them. This strategy failed when the consecutive
string was the result of a macro expansion.