[libc] Don't use _ctype symbol
The tradicional implementation of ctype uses the symbol _ctype,
(for example V7 or plan9 implementations), but _ctype is a
name that can be used by local variables, and they can hide the
symbol used by ctype.h. The best solution is to use the __
namespace, which cannot be used in any case.