[driver] Improve message error generation
We need the message errors of the linker because the user needs
them to catch symbol errors. The fprintf of spawn was only seen
with -d, so in case of missing some tool we were failing without
seeing any message. Changing _Exit to abort forces validatetools()
to print a message. In the same way, if the first tool that fails
if different of CC1 or LD then we have to print an internal error
message, because CC1 and LD are the only tools authorized to fail.