[libc] Add macro definitions for tolower() and toupper()
The library is not thread safe in a lot of different places, and it
doesn't try to be thread safe because it will require a more complex
design, out of the scope of this simple library. For this reason
it is not a problem to use a global static variable.