[as] Keep relative order of instruction with same verb
The order is important, because there are several ways to
implement an instruction, but not always they have the same
performance, so the table has to be built with an order
that ensure the correct seletion of instructions. The order
in sort of lines that are equal is unspecified, so we have
to add a workaround to ensure that the order is not
modified. This workaround basically insert a new field
that is the Number of record which can be used later
to enure the correct order.