Use + in recursive make actions
The rules that have the prefix + are executed always, even when make is
executed with -n. This prefix gives to make the information that this
line is going to execute a recursive make and it sets the correct
flags in MAKEFLAG for the children make, which will use the jobserver
of its parent.